Split test functions in folders
